Updates tests for downloads module
